requirements.txt
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
coverage==7.8.0
cryptography==44.0.3
docutils==0.21.2
gunicorn==23.0.0
id==1.5.0
idna==3.10
iniconfig==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
keyring==25.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
nh3==0.2.21
packaging==25.0
parse==1.20.2
pluggy==1.5.0
pycparser==2.22
Pygments==2.19.1
pytest==8.3.5
pytest-cov==6.1.1
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
requests-wsgi-adapter==0.4.1
rfc3986==2.0.0
rich==14.0.0
SecretStorage==3.3.3
setuptools==80.4.0
twine==6.1.0
urllib3==2.4.0
WebOb==1.8.9
wheel==0.45.1
whitenoise==6.9.0
